Refactor: Story/StoryPage/StoryCursor를 Domain/Model로 이동 + pbxproj UseCase 참조 정리
- Story, StoryPage, StoryCursor 타입을 Presentation → Domain/Model/StoryTypes.swift 이동
- Presentation/Story/Model/Story.swift에는 mock extension만 유지
- pbxproj에서 삭제된 UseCase/ PBXFileSystemSynchronizedRootGroup 참조 제거
- Domain → Presentation 역방향 의존 해소
